refactor(nft): implement NFT modal and collection components for enhanced asset management
