+++ apps/ui/src/modules/profile/components/nft/NftListMetadata.tsx
@@ -1,0 +1,33 @@
+import { Flex, Stack, Text } from '@chakra-ui/react';
+import { NftMetadataBlock } from './NftMetadataBlock';
+
+type Metadata = {
+  label: string;
+  value: string;
+};
+
+interface NftListMetadataProps {
+  metadata: Metadata[];
+}
+
+export const NftListMetadata = ({ metadata }: NftListMetadataProps) => {
+  return (
+    <Stack spacing={2}>
+      <Text>Metadata</Text>
+      <Flex gap={2} flexWrap="wrap">
+        {metadata.map((item) => (
+          <NftMetadataBlock
+            key={item.value}
+            title={item.label}
+            value={item.value}
+          />
+        ))}
+        {metadata.length === 0 && (
+          <Text fontSize="sm" color="grey.subtitle">
+            No metadata available.
+          </Text>
+        )}
+      </Flex>
+    </Stack>
+  );
+};
